restructure configs
