+++ betty/configs/__init__.py
@@ -1,0 +1,2 @@
+from .engine_dataclass import EngineConfig
+from .problem_dataclass import Config
